feat: refactored getPathfromId to folder service
